+++ packages/stream_chat_flutter/CHANGELOG.md
@@ -12,6 +12,10 @@
 
 - Animation curves changed from default `Curves.linear` to `Curves.easeOut` and `Curves.easeIn` for attachment controls.
 
+✅ Added
+
+- Added `MessageInput.customPortalOptions` property to add custom overlays to the `MessageInput`.
+
 ## 2.2.1
 
 ⚠️ Deprecated